+++ example/tst.dart
@@ -1,0 +1,7 @@
+import 'package:uuid/uuid.dart';
+
+void main() {
+  print(Uuid().v1());
+
+ // runApp(ng.AppComponentNgFactory);
+}